requirements.txt
coverage
fabric
jinja2
mock
nose
linode-python
pea
pycli
sphinx
-e git://github.com/lacostej/nosyd.git#egg=nosyd
